Updated eslint-config-shakacode to 13.2.0
Big change is that trailing commas are the default for function calls.

Autofixer makes this easy!

Be sure to run

npm run lint -- --fix

and you will be updated!

NOTE: Webpack config files must not have this set as of now, or else you
get a syntax error.

Webpack files need this, as they error if there's a trailing comma for
function calls!

/* eslint comma-dangle: ["error",
  {"functions": "never", "arrays": "only-multiline", "objects":
"only-multiline"} ] */